add: Implemented Google OAuth2 to access the user's email address
# Steps:
1. We first get a Google Auth url with the scope for acceessing a users
email address.  This will be expanded on later but we need a way to get
their address so we can setup IMAP.
2. Once the user has completed Google's authorization and returned a
code, we get their tokens and check their access_token info for their
email.
3. Now that we have this we can store the users refresh token in the
database so we can reauthenticate them later.  We will then need to
implement a second consent step where they give access to their email
account, not just the address.
4. For this we need to remember to generate a url using the parameter
include_granted_scopes: true.  This will mean we can use incremental
validation (google best practice) and hopefully means we will be able to
use the token we get the email with to also connect to IMAP.  This was
an issue previously.

# Additional Changes:
 - Changed the typescript ecmascript version to 2018 to allow promises.
 - Added the express-session library to add a random encoded 'state' to
the request to guard against CSRF attacks.